style

+++ src/components/DevLogCard/index.tsx
@@ -4,7 +4,6 @@ import {
   IonCardHeader,
   IonCardContent,
   IonCardTitle,
-  IonFabButton,
   IonIcon,
   IonRouterLink,
 } from "@ionic/react"
@@ -13,49 +12,67 @@ import { DevLog } from "../../util/types"
 import { dateFormatter, mdConverter } from "../../util"
 import { GetPWADetailUrl } from "../../routes"
 import { trash } from "ionicons/icons"
+import "./styles.css"
 
 interface ContainerProps {
   devLog: DevLog
-  onDelete: (logId: number) => void
+  isLinkable: boolean
+  onDelete?: (logId: number) => void
 }
 
-const DevLogCard: React.FC<ContainerProps> = ({ devLog, onDelete }) => {
+const DevLogCard: React.FC<ContainerProps> = ({
+  devLog,
+  isLinkable,
+  onDelete,
+}) => {
   const handleDelete = (e: any) => {
     e.preventDefault()
-    onDelete(devLog.logId)
+    onDelete && onDelete(devLog.logId)
   }
 
   return (
     <IonCard className="line-around">
-      <IonCardHeader>
-        {devLog.canDelete && (
-          <IonFabButton
-            className="CardFabButton"
-            style={{ zIndex: "100000" }}
-            size="small"
-            onClick={handleDelete}
-          >
-            <IonIcon color="danger" icon={trash} />
-          </IonFabButton>
-        )}
-        <IonCardTitle>
+      <IonCardHeader className="bottom-line-border" style={{ padding: "8px" }}>
+        <IonCardTitle className="DevLogCardTitle">
           <img src={devLog.icon} height="30px" width="30px" />
-          <IonRouterLink
-            style={{ paddingLeft: "8px" }}
-            routerLink={GetPWADetailUrl(devLog.appName)}
-          >
-            {devLog.appName}
-          </IonRouterLink>
+          {isLinkable ? (
+            <IonRouterLink
+              style={{ paddingLeft: "8px" }}
+              routerLink={GetPWADetailUrl(devLog.appName)}
+            >
+              {devLog.appName}
+            </IonRouterLink>
+          ) : (
+            <span style={{ paddingLeft: "8px" }}>{devLog.appName}</span>
+          )}
         </IonCardTitle>
       </IonCardHeader>
       <IonCardContent>
         <div
+          style={{ padding: "16px" }}
           dangerouslySetInnerHTML={{ __html: mdConverter.makeHtml(devLog.log) }}
         />
-        <div style={{ display: "flex", justifyContent: "flex-end" }}>
-          {dateFormatter(devLog.loggedAt)}
-        </div>
       </IonCardContent>
+      <div
+        style={{
+          display: "flex",
+          justifyContent: "space-between",
+          alignItems: "center",
+          padding: "8px"
+        }}
+      >
+        <div>
+          {devLog.canDelete && (
+            <button
+              style={{ background: "transparent" }}
+              onClick={handleDelete}
+            >
+              <IonIcon size="small" className="sub-color" icon={trash} />
+            </button>
+          )}
+        </div>
+        <div>{dateFormatter(devLog.loggedAt)}</div>
+      </div>
     </IonCard>
   )
 }
